Issue#124: Refactored the Github project structure according to the clean_code suggestions from the code review. Crawler is still runing after the changes
