+++ src/crawler/exceptions/unknown_exception.py
