+++ src/crawler/spider/spider.py
@@ -1,7 +1,7 @@
 import requests
 
 def executeRequest(url, header):
-    """Methode erhält eine URL und einen header_creater als String übergeben.
+    """Methode erhält eine URL und einen header als String übergeben.
     Stellt damit eine Request an besagte URL und liefert den HTML Inhalt der Response als String zurück."""
 
     response = requests.get(url, headers=header)
