+++ tests/test_config_reader.py
@@ -1,6 +1,5 @@
 import yaml
 
-import crawler.config_reader.config_reader
 from config_reader import config_reader
 import os
 import unittest
@@ -15,10 +14,10 @@ class TestConfigReader(unittest.TestCase):
         file.close()
         test_read = config_reader.read_settings_file("./test_settings.yaml")
         os.remove("test_settings.yaml")
-        # assert test_read["client"] == "safari", "Error in config_reader method read_settings_file. Expected value for " \
+        # assert test_read["client"] == "safari", "Error in config method read_settings_file. Expected value for " \
         #                                         "client is safari. "
         self.assertEqual(test_read["client"], "safari",
-                         "Error in config_reader method read_settings_file. Expected value for " \
+                         "Error in config method read_settings_file. Expected value for " \
                          "client is safari. ")
 
     def test_url_reader(self):
@@ -29,15 +28,15 @@ class TestConfigReader(unittest.TestCase):
         test_read = config_reader.read_settings_file("./test_urls.yaml")
         os.remove("test_urls.yaml")
         # assert test_read[
-        #            1] == "https://tagesschau.de", "Error in config_reader method read_url_list. Expected value for " \
+        #            1] == "https://tagesschau.de", "Error in config method read_url_list. Expected value for " \
         #                                           "list at index 1 is https://tagesschau.de. "
         self.assertEqual(test_read[1], "https://tagesschau.de",
-                         "Error in config_reader method read_url_list. Expected value for " \
+                         "Error in config method read_url_list. Expected value for " \
                          "list at index 1 is https://tagesschau.de. ")
 
 
     def test_settings_reader_aws(self):
-        test_read = config_reader.read_config('../input/url.yaml', '../input/settings.yaml')
+        test_read = config_reader.read_config('../config/url.yaml', '../config/settings.yaml')
         self.assertEqual(test_read["aws_env"], False,
-                         "Error in config_reader method read_settings_file. Expected value for " \
+                         "Error in config method read_settings_file. Expected value for " \
                          "aws_env is not 'False'. ")
